+++ util/db_test_util.cc
@@ -28,6 +28,7 @@ SpecialEnv::SpecialEnv(Env* base)
   manifest_sync_error_.store(false, std::memory_order_release);
   manifest_write_error_.store(false, std::memory_order_release);
   log_write_error_.store(false, std::memory_order_release);
+  random_file_open_counter_.store(0, std::memory_order_relaxed);
   log_write_slowdown_ = 0;
   bytes_written_ = 0;
   sync_counter_ = 0;